final lab, integrate weather api

+++ producer/src/main/java/miu/edu/bdt/producer/dto/Current.java
@@ -18,19 +18,20 @@ public class Current {
     private Condition condition;
     private float wind_mph;
     private float wind_kph;
-    private int wind_degree;
+    private float wind_degree;
     private String wind_dir;
-    private int pressure_mb;
+    private float pressure_mb;
     private float pressure_in;
     private float precip_mm;
     private float precip_in;
-    private int humidity;
-    private int cloud;
+    private float humidity;
+    private float cloud;
     private float feelslike_c;
     private float feelslike_f;
-    private int vis_km;
-    private int vis_miles;
-    private int uv;
+    private float vis_km;
+    private float vis_miles;
+    private float uv;
     private float gust_mph;
     private float gust_kph;
+
 }
